Adding Dockerfile

+++ cvm/views.py
@@ -1,3 +1,5 @@
+import os
+
 import pdfkit
 from django.http import HttpResponse
 from django.shortcuts import render, redirect
@@ -5,7 +7,7 @@ from django.template import loader
 
 from .forms import RegisterForm
 from .models import CV
-from .utils import save_resume
+from .utils import save_resume, find_cv
 
 
 def create_resume(request):
@@ -24,7 +26,9 @@ def create_resume(request):
 def view_resume(request, id):
     if not request.user.is_authenticated:
         return redirect('login')
-    cv = CV.objects.get(pk=id)
+    cv = find_cv(id)
+    if not cv:
+        return redirect('user-resume')
     return render(request, 'view_resume.html',
                   {'cv': cv, 'edu': cv.education.all(), 'exp': cv.experience.all()})
 
@@ -32,8 +36,11 @@ def view_resume(request, id):
 def update_resume(request, id):
     if not request.user.is_authenticated:
         return redirect('login')
+    new_cv = find_cv(id)
+    if not new_cv:
+        return redirect('user-resume')
     _cv = CV.objects.filter(pk=id)
-    new_cv = _cv.first()
+
     if new_cv.user.pk != request.user.pk:
         return redirect('create-resume')
 
@@ -49,7 +56,9 @@ def update_resume(request, id):
 def print_resume(request, id):
     if not request.user.is_authenticated:
         return redirect('login')
-    new_cv = CV.objects.get(pk=id)
+    new_cv = find_cv(id)
+    if not new_cv:
+        return redirect('user-resume')
     if new_cv.user.pk != request.user.pk:
         return redirect('create-resume')
     template = loader.get_template('print_resume.html')
@@ -59,7 +68,7 @@ def print_resume(request, id):
         'encoding': 'UTF-8',
         'enable-local-file-access': None
     }
-    css = '/home/shivansh/PycharmProjects/CV_Maker/cvm/static/css/cv.css'
+    css = os.path.abspath(os.getcwd()) + '/cvm/static/css/cv.css'
     pdf = pdfkit.from_string(html, False, option, css=css)
     response = HttpResponse(pdf, content_type='application/pdf')
     response['Content-Disposition'] = 'attachments;filename=resume.pdf;'
@@ -67,6 +76,8 @@ def print_resume(request, id):
 
 
 def user_resume(request):
+    if request.user.is_anonymous:
+        return redirect('login')
     cv = request.user.cv.all().order_by('-created_at').values()
     context = {"cvs": cv}
     return render(request, 'user_resume.html', context)
@@ -75,7 +86,9 @@ def user_resume(request):
 def delete_resume(request, id):
     if not request.user.is_authenticated:
         return redirect('login')
-    cv = CV.objects.get(pk=id)
+    cv = find_cv(id)
+    if not cv:
+        return redirect('user-resume')
     if cv.user.pk != request.user.pk:
         return redirect('create-resume')
     cv.delete()
